requirements.txt
# ──────────────────────────────────────────────────────────────
# Traffic Sign Recognition — Python Dependencies
# ──────────────────────────────────────────────────────────────

# Deep Learning
tensorflow>=2.10.0

# Image Processing
opencv-python>=4.5.0
Pillow>=9.0.0

# Data Science
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=1.0.0

# Visualization
matplotlib>=3.5.0
seaborn>=0.11.0

# Web Application
flask>=2.0.0
werkzeug>=2.0.0
